Update effects.h
I've made two new effects that work with the rest of the effects in this sketch:
First new effect: ShadesOutline
Leds start lighting in order starting from the upper left of the matrix (right eye), and outline the shades in a clockwise manner. When an outline finishes, the matrix is cleared, a new random color is selected, and the outline starts over. One idea that I haven't implemented is to have the outline travel more slowly when moving vertically because there are only 5 pixels to traverse, similar to the code for the color wipe effect.
Second new effect: hearts
One animated heart on each eye that starts 2 pixels high and grows to 5 pixels high. The hearts start out pale pink. At each step they get darker until they are fully red when fully grown. The fully grown heart displays for slightly longer than the other ones, then the matrix is cleared and the animation repeats.